Remove Flash_write_reserve related functions. Move some functions from ruuvi_task_flash_ringbuffer to ruuvi_task_flashdb.
