+++ src/ruuvi.drivers.c/src/tasks/ruuvi_task_flash_ringbuffer.c
@@ -14,7 +14,6 @@
 #include "ruuvi_task_flash_ringbuffer.h"
 #include "ruuvi_task_flashdb.h"
 #include "fds.h"
-#include "macronix_flash.h"
 #include <string.h>
 
 
@@ -51,17 +50,11 @@ static inline void LOGDf (const char * const msg, ...)
 
 /* TSDB object */
 static struct fdb_tsdb tsdb;
-static rd_status_t is_macronix_present = RD_ERROR_NOT_INITIALIZED;
 
 rd_status_t rt_flash_ringbuffer_create (const char *partition, fdb_get_time get_time, const bool format_db)
 {
-  bool with_ext_flash = (rt_flash_ringbuffer_ext_flash_exists()==RD_SUCCESS);
-
   // change to high performance mode during flashDB initialization for quicker setup
-  bool high_power =1;
-  if(with_ext_flash) {
-    mx_high_performance_switch (high_power);
-  }
+  rt_macronix_high_performance_switch(true);
 
   /* Time Series database initialization
    */
@@ -82,10 +75,7 @@ rd_status_t rt_flash_ringbuffer_create (const char *partition, fdb_get_time get_
   }
   
   // change to low power mode after flashDB initialization for quicker setup
-  high_power =0;
-  if(with_ext_flash) {
-    mx_high_performance_switch (high_power);
-  }
+  rt_macronix_high_performance_switch(false);
 
   return rt_flashdb_to_ruuvi_error(result);
 }
@@ -113,22 +103,14 @@ void rt_flash_ringbuffer_read (const fdb_tsl_cb callback, const ri_comm_xfer_fp_
 
 rd_status_t rt_flash_ringbuffer_clear (void) {
 
-  bool with_ext_flash = (rt_flash_ringbuffer_ext_flash_exists()==RD_SUCCESS);
-
   // change to high performance mode during flashDB initialization for quicker setup
-  bool high_power =1;
-  if(with_ext_flash) {
-    mx_high_performance_switch (high_power);
-  }
+  rt_macronix_high_performance_switch(true);
 
   fdb_tsl_clean(&tsdb);
   LOGD("Ringbuffer cleared\r\n");
 
   // change to low power mode after flashDB initialization for quicker setup
-  high_power =0;
-  if(with_ext_flash) {
-    mx_high_performance_switch (high_power);
-  }
+  rt_macronix_high_performance_switch(false);
 
   return RD_SUCCESS;
 }
@@ -170,64 +152,4 @@ rd_status_t rt_flash_ringbuffer_statistic (uint8_t* const statistik) {
   return err_code;
 }
 
-void rt_print_flash_statistic(void) {
-
-  rd_status_t err_code = RD_SUCCESS;
-      
-  // gather flash statistics
-  fds_stat_t stat;
-  fds_stat(&stat);
-
-  LOGDf("Open records %d \r\n", stat.open_records);
-  LOGDf("Valid records %d \r\n", stat.valid_records);
-  LOGDf("Dirty records %d \r\n", stat.dirty_records);
-  LOGDf("Words reserved %d \r\n", stat.words_reserved);
-  LOGDf("Words used %d \r\n", stat.words_used);
-  LOGDf("Words available %d \r\n", stat.pages_available * FDS_VIRTUAL_PAGE_SIZE);
-  LOGDf("Largest contig %d \r\n", stat.largest_contig);
-  LOGDf("Freeable words %d \r\n", stat.freeable_words);
-  LOGDf("Filesystem corruption %d \r\n", stat.corruption);
-  
-}
-
-rd_status_t rt_flash_ringbuffer_ext_flash_exists(void) {
-
-  if(is_macronix_present == RD_ERROR_NOT_INITIALIZED) {
-    uint8_t manufacturer_id;
-    uint8_t device_id;
-
-    LOGD("Checking external flash existence\r\n");
-
-    while(mx_busy() == RD_ERROR_BUSY){
-      ri_yield();
-    }
-
-    rd_status_t err_code = mx_read_rems(&manufacturer_id, &device_id);
-
-    if(err_code==RD_SUCCESS) {
-      LOGDf("manufacturer_id %x, device_id %x\r\n", manufacturer_id, device_id);
-
-      if(manufacturer_id==0xc2 && device_id==0x17) {      
-        is_macronix_present = RD_SUCCESS;
-      } else {
-        is_macronix_present = RD_ERROR_NOT_FOUND;
-      }
-    }
-  }
-  return is_macronix_present;
-}
-
-rd_status_t rt_reset_macronix_flash(void) {
-  rd_status_t err_code;
-  err_code = RD_SUCCESS;
-  if(rt_flash_ringbuffer_ext_flash_exists()==RD_SUCCESS) {
-    LOGDf("Triggering Macronix flash chip erase\r\n");
-    err_code |= mx_high_performance_switch(true);
-    mx_spi_ready_for_transfer();
-    err_code |= mx_chip_erase();
-  }
-  return err_code;
-}
-
 #endif
-
