+++ src/ruuvi.drivers.c/src/tasks/ruuvi_task_flash_ringbuffer.h
@@ -65,10 +65,6 @@ rd_status_t rt_flash_ringbuffer_statistic (
   uint8_t* const statistik
 );
 
-rd_status_t rt_flash_ringbuffer_ext_flash_exists(void);
-
-rd_status_t rt_reset_macronix_flash(void);
-
 #endif // RUUVI_TASK_FLASH_RINGBUFFER_H
 
 #endif
